fix(aries-components): ensure ConnectionRecord is only added once
When a new ConnectionRecord is created it will be added to the store by
the EventListener. We previously also added it in de reducer of the
ConnectionThunks.createConnection() thunk, resulting in duplicate
ConnectionRecords in the store. This has now been removed, so only the EventListener
is adding it to the store.

Signed-off-by: Karim Stekelenburg <karim@animo.id>