refactor: change ConfigurationContext for supportedLanguages
The ConfigurationContext still has a supportedLanguage prop that now is
of type Locales[].

It's also possible now pass a defaultLanguage (optional) in the initLanguage
function. It was always English by default before. The languages array in the
Language screen is not hardcoded anymore and in the Settings Screen the
currentLanguage doesn't use anymore the hardcoded languages array.

In the supportedLanguages prop from ConfigurationContext you can pass
the locales you want to support (ex: [Locales.en, Locales.fr])

Eventually it would be a nice feature to add where in the Settings
screen if the wallet you are working on supports only one language (ex:
Portuguese), it would make sense to not display in the App Setttins the
Language option to switch the (Only show the Biometrics and Change PIN).

Signed-off-by: fc-santos <filipesantos@live.ca>